Move to using a root command class
