+++ lib/jay/cli/root_command.rb
@@ -1,0 +1,18 @@
+module Jay
+  module Cli
+    class RootCommand < Thor
+      desc "version", "Print the version"
+      def version
+        say Jay::VERSION
+      end
+
+      def self.basename
+        "jay"
+      end
+
+      def self.exit_on_failure?
+        true
+      end
+    end
+  end
+end
